requirements.txt
absl-py==0.11.0
albumentations==0.5.2
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.5.1
async-generator==1.10
atpublic==2.1.3
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cycler==0.10.0
Cython==0.29.21
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.7.1
dictdiffer==0.8.1
distro==1.5.0
dpath==2.0.1
dulwich==0.20.19
dvc==1.11.16
efficientnet-pytorch==0.7.0
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.9.0
flatten-dict==0.3.0
flufl.lock==3.2
ftfy==5.9
funcy==1.15
future==0.18.2
gitdb==4.0.5
GitPython==3.1.13
grandalf==0.6
grpcio==1.35.0
idna==2.10
imagecodecs==2021.2.26
imageio==2.9.0
imgaug==0.4.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kneed==0.7.0
lazy-object-proxy==1.5.2
loguru==0.5.3
lxml==4.6.2
mailchecker==4.0.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
nanotime==0.5.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nd2reader==3.2.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathlib2==2.3.5
pathspec==0.8.1
pexpect==4.8.0
phonenumbers==8.12.18
pickleshare==0.7.5
Pillow==8.1.0
PIMS==0.5
plotly==4.14.3
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pycparser==2.20
pydot==1.4.2
Pygments==2.8.0
pygtrie==2.3.2
pylint==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-benedict==0.23.2
python-dateutil==2.8.1
python-fsutil==0.4.0
python-slugify==4.0.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rich==9.11.0
rsa==4.7.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
Shapely==1.7.1
shortuuid==1.0.1
shtab==1.3.4
six==1.15.0
slicerator==1.0.0
smmap==3.0.5
soupsieve==2.2
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.3
testpath==0.4.4
text-unidecode==1.3
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2021.2.26
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
voluptuous==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
zc.lockfile==2.0
